Rendertextures & small graphics improvements
1. naive implementation of rendertexture: object that contains a handle
to an FBO and handles to textures attached to the color and depth
buffers. Currently fixed to only 1 FBO. Generalizing this in another
update.
2. reversed uv values on quad and flipped texture data loaded from disk.
- After rendering the output of an FBO texture I finally realized I had
been rendering textures upside down.
3. wrote enableExtensions and enableExtension functions to handle
enabling openGL functionality that is not part of the current webGL
specification
4. Began writing the generalized implemenation of rendertexture.
Rendertexture hashtable, new xml fileformat .rtex, writing
loadRenderTexture function